Remove Text, add fromString
The current formatString is based on Read, as it allows us to avoid
brinding in dependencies for parsing text (e.g. Megaparsec). This means
we are restricted to String, unfortunately. As such, there is little point
in having our format functions return Text; might as well use String.

Thus we can get rid of Text entirely, until we come up with a better way
to parse Text that doesn't require significantly increasing our dependency
footprint.